debugging with discord gang
