+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "wxer_lib"
-version = "0.11.1"
+version = "0.12.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
@@ -8,7 +8,7 @@ edition = "2021"
 [dependencies]
 anyhow = "1.0.75"
 chrono = { version = "0.4", features = ["serde"] }
-chrono-tz = "0.9.0"
+chrono-tz = { version = "0.10.0", features = ["serde"] }
 csv = "1.3.0"
 derive_more = "0.99.17"
 futures = "0.3.30"
