+++ src/fetch/rpi_station.rs
@@ -1,7 +1,9 @@
 use std::collections::BTreeMap;
+use std::sync::Arc;
 
 use chrono::{DateTime, Duration, Timelike, Utc};
 use anyhow::Result;
+use chrono_tz::US::Eastern;
 use serde::Deserialize;
 
 use crate::*;
@@ -14,14 +16,14 @@ use crate::Layer::*;
 
 // station URL is going to be something like http://rpi_address:8000
 // todo: better name for rpi station
-pub async fn import(station_url: &str, date: DateTime<Utc>) -> Result<db::StationData> {
+pub async fn import(station_url: &str, date: DateTime<Utc>, station: Arc<Station>) -> Result<db::StationData> {
 
-    let station_data_url = format!("{station_url}/location.json");
+    // let station_data_url = format!("{station_url}/location.json");
 
     // use this later
-    let station = reqwest::get(station_data_url)
-        .await?
-        .text().await?;
+    // let station = reqwest::get(station_data_url)
+    //     .await?
+    //     .text().await?;
 
     let url = format!("{station_url}/{}.csv", date.format("%Y-%m-%d").to_string());
     let resp: String = reqwest::get(url)
@@ -59,7 +61,8 @@ fn try_parse_entry(record: Result<RaspPiEntry, csv::Error>) -> Result<(DateTime<
     let station: Station = Station { 
         name: "RPI@ValentineHouse".into(), 
         altitude: Altitude::new(4.8, Meter), 
-        coords: (43.13, -70.92), 
+        coords: (43.13, -70.92).into(), 
+        time_zone: Eastern,
     };
 
     let mut wx = HashMapWx::new(dt, station);
